requirements.txt
Flask==3.0.2
flask_cors>=4.0.0
mysqlclient
pymysql
numpy>=1.26.4
pandas>=2.2.0
pymongo>=4.6.1
python-dotenv>=1.0.1
Jinja2>=3.1.2
python-dateutil>=2.8.2
Werkzeug>=3.0.0
cryptography>=3.4.8
moviepy
wget
boto3
gevent
gevent-websocket